+++ changelog.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## Updated
+## Added
+- Added the `newAggregate()` function to customize the aggregate query builder. ([#11](https://github.com/Jeroen-G/Explorer/pull/11))
+
+## Changed
 - Updated dev dependency Infection to 0.20
 
 ## [1.0.0]